Cleared local repo, modified gemfiles. Added 2 files to complete first task of ruby-automation cource: Main_project_Helpers.rb contains helpers-methods. Main_project_Tests_TestUnit.rb contains tests to check demo.redmine.org
