+++ .gitignore
@@ -9,6 +9,8 @@
 /test/tmp/
 /test/version_tmp/
 /tmp/
+Gemfile.lock
+.ruby-version
 
 # Used by dotenv library to load environment variables.
 # .env
