+++ Gemfile
@@ -5,3 +5,6 @@ source "https://rubygems.org"
 # gem "rails"
 gem 'selenium-webdriver', '~> 2.53', '>= 2.53.3'
 gem "test-unit"
+gem "faker", '~> 1.7', '>= 1.7.3'
+gem "rspec"
+gem 'cucumber', '~> 2.4'
